+++ search/src/negamax/utils.rs
@@ -3,12 +3,12 @@ use evaluation::piece_value;
 use uci::commands::Score;
 
 #[inline(always)]
-pub fn see_naive(board: &Board, capture_move: ChessMove) -> i16 {
+pub fn see_naive(board: &Board, capture_move: ChessMove) -> i32 {
     if let (Some(captured_piece), Some(capturing_piece)) = (
         board.piece_on(capture_move.get_dest()),
         board.piece_on(capture_move.get_source()),
     ) {
-        (piece_value(captured_piece) - piece_value(capturing_piece)) as i16
+        (piece_value(captured_piece) - piece_value(capturing_piece)) as i32
     } else {
         0
     }
@@ -28,8 +28,8 @@ pub fn calculate_dynamic_lmr_reduction(depth: u64, move_index: usize, score: i32
 }
 
 #[inline(always)]
-pub fn convert_mate_score(score: i16, pv: &Vec<ChessMove>) -> Score {
-    let mate_in = (pv.len() as i16 + 1) / 2;
+pub fn convert_mate_score(score: i32, pv: &Vec<ChessMove>) -> Score {
+    let mate_in = (pv.len() as i32 + 1) / 2;
 
     if score > 0 {
         Score::Mate(mate_in)
@@ -39,6 +39,6 @@ pub fn convert_mate_score(score: i16, pv: &Vec<ChessMove>) -> Score {
 }
 
 #[inline(always)]
-pub fn convert_centipawn_score(score: i16) -> Score {
+pub fn convert_centipawn_score(score: i32) -> Score {
     Score::Centipawns(score)
 }
